feat(issue-425): refactor tree structure for NASA standards compliance
- Remove redundant app prefixes (ClientHostCli → Cli, ServerArgs → Args)
- Fix file naming to strict PascalCase for C++ files
- Organize files per Loi de Proximité (headers + impl in same folder)
- Verify folder depth (max 2 levels, exception for cuda/fragments)
- Update includes, CMake, and documentation
- Zero compiler warnings, fully deterministic
- Complies with NPR-7150.2D + NASA-STD-8739.8B
