+++ apps/headless/src/main.cu
@@ -1,5 +1,5 @@
 /*
- * @file apps/headless/main.cu
+ * @file apps/headless/src/main.cu
  * @author Luis1454
  * @project BLITZAR
  * @brief Application entry points and host executables for BLITZAR.
