+++ docs/quality/manifest/evidence.json
@@ -12,7 +12,7 @@
         "EVD_GITHUB_PR_TEMPLATE": ".github/PULL_REQUEST_TEMPLATE.md",
         "EVD_AGENTS": "AGENTS.md",
         "EVD_ROOT_CMAKELISTS": "CMakeLists.txt",
-        "EVD_APP_LAUNCHER_MAIN": "apps/launcher/main.cpp",
+        "EVD_APP_LAUNCHER_MAIN": "apps/launcher/src/main.cpp",
         "EVD_DOC_SERVER_PROTOCOL": "docs/server-protocol.md",
         "EVD_DOC_SERVER_PROTOCOL_SCHEMA": "runtime/protocol/server-protocol-schema.json",
         "EVD_DOC_WEB_GATEWAY": "docs/web-gateway.md",
