+++ tests/cmake/targets_gtests.cmake
@@ -79,10 +79,10 @@ if(BLITZAR_TEST_UNIT_MODULE_SOURCES)
             "${BLITZAR_ROOT_DIR}/modules/cli/module_cli_server_ops.cpp"
             "${BLITZAR_ROOT_DIR}/modules/cli/module_cli_commands.cpp"
             ${BLITZAR_RUNTIME_COMMAND_SOURCES}
-            "${BLITZAR_ROOT_DIR}/apps/client-host/client_host_cli.cpp"
-            "${BLITZAR_ROOT_DIR}/apps/client-host/client_host_cli_args.cpp"
-            "${BLITZAR_ROOT_DIR}/apps/client-host/client_host_module_ops.cpp"
-            "${BLITZAR_ROOT_DIR}/apps/client-host/client_host_cli_text.cpp"
+            "${BLITZAR_ROOT_DIR}/apps/client-host/src/Cli.cpp"
+            "${BLITZAR_ROOT_DIR}/apps/client-host/src/CliArgs.cpp"
+            "${BLITZAR_ROOT_DIR}/apps/client-host/src/ModuleOps.cpp"
+            "${BLITZAR_ROOT_DIR}/apps/client-host/src/CliText.cpp"
             "${BLITZAR_ROOT_DIR}/engine/src/config/SimulationArgsParse.cpp"
             "${BLITZAR_ROOT_DIR}/engine/src/config/SimulationOptionRegistry.cpp"
             "${BLITZAR_ROOT_DIR}/engine/src/config/SimulationOptionRegistryApply.cpp"
